feat: refactor runes info tab
